Text positioning improvement
The old code relied on a fixed x,y coordinate on the screen.  This would
mean that the image would only appear correctly on a single resolution.

This patch is the first step towards allowing things to dynamically change
to fit onto different monitor sizes.

+++ jt_text.h
@@ -5,9 +5,15 @@ typedef struct jt_text_struct {
     /* Configuration */
     char *text;
     uint32_t length;
-    unsigned int scale;
-    unsigned int x_position;
-    unsigned int y_position;
+    /* TODO: Change definition of scale to represent the number of
+     * characters that we want to fit across the screen */
+    uint32_t scale;
+    /* Position is stored as a value between 0.0 and 1.0.
+     * 0.0 represents left-aligned text.
+     * 0.5 represents centred text.
+     * 1.0 represents right-alligned text. */
+    float x_position;
+    float y_position;
 
     SDL_Texture *typed_texture;
     SDL_Texture *untyped_texture;
